Fixed bugs
Apparently, I do not know how to spell the word "Autonomous." That broke the code when trying to build.

I also apparently left off a `}` somewhere. So I just added that and it made the compiler happy.

Apparently, the "Autonomous" class does not exist, so all the code extends the `LinearOpMode` class now.